CurveAlign ROI statistics flag check
This gives a value to the parameter ROIstatsFLAG, which is used to check
whether or not the statistics from a particular ROI are there or not.
If it is empty, the code procedes.

Without this an empty ROI would halt the program.  This is particularly
relevant for tiling procedures, where there may be many empty ROIs
because of the automatic division and selection.